# lib/api-zod/src/generated/types/limitsConfig.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface LimitsConfig {
  max_records?: number;
  max_per_source?: number;
}