# lib/api-zod/src/generated/types/runListItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { RunListItemStatus } from "./runListItemStatus";
import type { RunMetrics } from "./runMetrics";
import type { StageMetrics } from "./stageMetrics";

export interface RunListItem {
  run_id: string;
  run_name: string;
  status: RunListItemStatus;
  created_at: string;
  updated_at: string;
  metrics?: RunMetrics;
  stage_metrics?: StageMetrics[];
}