# lib/api-zod/src/generated/types/runMetrics.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface RunMetrics {
  ingest_success_rate?: number;
  dedup_ratio?: number;
  avg_chunk_tokens?: number;
  generation_yield?: number;
  validation_pass_rate?: number;
  avg_final_score?: number;
  total_records?: number;
  drop_count?: number;
}