# lib/api-zod/src/generated/types/stageMetrics.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface StageMetrics {
  stage?: string;
  input_count?: number;
  output_count?: number;
  latency_ms?: number;
  notes?: string;
}